+++ libs/langchain-core/src/messages/base.ts
@@ -68,7 +68,7 @@ export interface FunctionCall {
 
 export type BaseMessageFields<
   TStructure extends MessageStructure = MessageStructure,
-  TRole extends MessageType = MessageType
+  TRole extends MessageType = MessageType,
 > = Pick<Message, "id" | "name"> & {
   content?: $InferMessageContent<TStructure, TRole>;
   contentBlocks?: Array<ContentBlock.Standard>;
@@ -197,7 +197,7 @@ function stringifyWithDepthLimit(obj: any, depthLimit: number): string {
  */
 export abstract class BaseMessage<
     TStructure extends MessageStructure = MessageStructure,
-    TRole extends MessageType = MessageType
+    TRole extends MessageType = MessageType,
   >
   extends Serializable
   implements Message<TStructure, TRole>
@@ -262,7 +262,9 @@ export abstract class BaseMessage<
       | BaseMessageFields<TStructure, TRole>
   ) {
     const fields: BaseMessageFields<TStructure, TRole> =
-      typeof arg === "string" || Array.isArray(arg) ? { content: arg } : arg;
+      typeof arg === "string" || Array.isArray(arg)
+        ? ({ content: arg } as BaseMessageFields<TStructure, TRole>)
+        : arg;
     if (!fields.additional_kwargs) {
       fields.additional_kwargs = {};
     }
@@ -422,11 +424,17 @@ export function isOpenAIToolCallArray(
 
 export function _mergeDicts(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  left: Record<string, any> = {},
+  left: Record<string, any> | undefined,
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  right: Record<string, any> = {}
+  right: Record<string, any> | undefined
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-): Record<string, any> {
+): Record<string, any> | undefined {
+  if (left === undefined && right === undefined) {
+    return undefined;
+  }
+  if (left === undefined || right === undefined) {
+    return left ?? right;
+  }
   const merged = { ...left };
   for (const [key, value] of Object.entries(right)) {
     if (merged[key] == null) {
@@ -454,6 +462,8 @@ export function _mergeDicts(
       } else {
         merged[key] += value;
       }
+    } else if (typeof merged[key] === "number") {
+      merged[key] = merged[key] + value;
     } else if (typeof merged[key] === "object" && !Array.isArray(merged[key])) {
       merged[key] = _mergeDicts(merged[key], value);
     } else if (Array.isArray(merged[key])) {
@@ -568,7 +578,7 @@ export function _mergeObj<T = any>(
  */
 export abstract class BaseMessageChunk<
   TStructure extends MessageStructure = MessageStructure,
-  TRole extends MessageType = MessageType
+  TRole extends MessageType = MessageType,
 > extends BaseMessage<TStructure, TRole> {
   abstract concat(chunk: BaseMessageChunk): BaseMessageChunk<TStructure, TRole>;