+++ libs/providers/langchain-xai/src/chat_models/completions.ts
@@ -890,7 +890,7 @@ export class ChatXAI extends ChatOpenAICompletions<ChatXAICallOptions> {
 
   override withStructuredOutput<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    RunOutput extends Record<string, any> = Record<string, any>
+    RunOutput extends Record<string, any> = Record<string, any>,
   >(
     outputSchema:
       | InteropZodType<RunOutput>
@@ -901,7 +901,7 @@ export class ChatXAI extends ChatOpenAICompletions<ChatXAICallOptions> {
 
   override withStructuredOutput<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    RunOutput extends Record<string, any> = Record<string, any>
+    RunOutput extends Record<string, any> = Record<string, any>,
   >(
     outputSchema:
       | InteropZodType<RunOutput>
@@ -912,7 +912,7 @@ export class ChatXAI extends ChatOpenAICompletions<ChatXAICallOptions> {
 
   override withStructuredOutput<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    RunOutput extends Record<string, any> = Record<string, any>
+    RunOutput extends Record<string, any> = Record<string, any>,
   >(
     outputSchema:
       | InteropZodType<RunOutput>
@@ -925,7 +925,7 @@ export class ChatXAI extends ChatOpenAICompletions<ChatXAICallOptions> {
 
   override withStructuredOutput<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    RunOutput extends Record<string, any> = Record<string, any>
+    RunOutput extends Record<string, any> = Record<string, any>,
   >(
     outputSchema:
       | InteropZodType<RunOutput>